Package rename and cleanup
