+++ go/lib/pktcls/class.go
@@ -12,9 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package class
+package pktcls
 
-import "encoding/json"
+import (
+	"encoding/json"
+)
 
 var (
 	_ json.Marshaler   = (*Class)(nil)
